servers/tradingview-mcp/pyproject.toml
[project]
name = "tradingview-mcp"
version = "0.1.0"
description = "Model Context Protocol server exposing crypto screener tools (TradingView Screener)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "mcp[cli]>=1.12.0",
  "pydantic>=2.0.0",
  "pydantic-settings>=2.0.0",
  "tradingview-screener>=0.6.4",
  "tradingview-ta>=3.3.0",
  "numpy>=1.24.0",
  "pandas>=2.0.0",
  "ta-lib>=0.4.0",
  "scipy>=1.10.0",
  "httpx>=0.24.0",
  "redis>=5.0.0",
  "asyncpg>=0.28.0",
  "pyyaml>=6.0",
  "fastapi>=0.104.0",
  "uvicorn>=0.24.0",
  "fastapi-mcp>=0.1.0",
]

[project.scripts]
# Entry point now points to packaged module
tradingview-mcp = "tradingview_mcp.server:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
tradingview_mcp = ["coinlist/*.txt", "configs/*.yaml"]

[tool.uv]
dev-dependencies = []
package = true

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"
